# Adapters/OpenApi/openapi.yaml
openapi: 3.0.4
info:
  title: CLI
  version: '1.0'
paths:
...
  '/api/roles/{roleName}/users':
    get:
      tags:
        - RoleUsers
      operationId: GetRoleUsersV1
      parameters:
        - name: roleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RoleUsersResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/RoleUsersResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/RoleUsersResponse'
      security:
        - Bearer: [ ]
  /api/v1/sensors:
    get:
      tags:
        - Sensors
      operationId: GetAllSensorsV1
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorResponse'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorResponse'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorResponse'
      security:
        - Bearer: [ ]
    post:
      tags:
        - Sensors
      operationId: CreateSensorV1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SensorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
  '/api/v1/sensors/{id}/readings':
    get:
      tags:
        - Sensors
      operationId: GetSensorReadingsV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SensorReadingResponsePagedResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/SensorReadingResponsePagedResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/SensorReadingResponsePagedResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
  '/api/v1/sensors/{id}':
    get:
      tags:
        - Sensors
      operationId: GetSensorByIdV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SensorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    put:
      tags:
        - Sensors
      operationId: UpdateSensorV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SensorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
    delete:
      tags:
        - Sensors
      operationId: DeleteSensorV1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - Bearer: [ ]
...
  '/api/users/{userId}/roles':
    get:
      tags:
        - UserRoles
      operationId: GetUserRolesV1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UserRolesResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/UserRolesResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/UserRolesResponse'
      security:
        - Bearer: [ ]
    post:
      tags:
        - UserRoles
      operationId: AssignUserRolesV1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesToUserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AssignRolesToUserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AssignRolesToUserRequest'
      responses:
        '200':
          description: OK
      security:
        - Bearer: [ ]
    delete:
      tags:
        - UserRoles
      operationId: RemoveUserRolesV1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveRolesFromUserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RemoveRolesFromUserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RemoveRolesFromUserRequest'
      responses:
        '200':
          description: OK
      security:
        - Bearer: [ ]
    put:
      tags:
        - UserRoles
      operationId: ReplaceUserRolesV1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesToUserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AssignRolesToUserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AssignRolesToUserRequest'
      responses:
        '200':
          description: OK
      security:
        - Bearer: [ ]
components:
  schemas:
...
    AssignRolesToUserRequest:
      required:
        - roles
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    RemoveRolesFromUserRequest:
      required:
        - roles
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
...
    RoleUsersResponse:
      required:
        - roleName
        - userIds
      type: object
      properties:
        roleName:
          type: string
          nullable: true
        userIds:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
...
    SensorReadingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sensorId:
          type: string
          format: uuid
        value:
          type: number
          format: double
        timestampUtc:
          type: string
          format: date-time
        unit:
          $ref: '#/components/schemas/Unit'
        organizationId:
          type: string
          format: uuid
        ingestionRunId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    SensorReadingResponsePagedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SensorReadingResponse'
          nullable: true
        totalCount:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
    SensorRequest:
      type: object
      properties:
        serial:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/SensorStatus'
        equipmentId:
          type: string
          format: uuid
        sensorTypeId:
          type: string
          format: uuid
        thresholdId:
          type: string
          format: uuid
      additionalProperties: false
    SensorResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        serial:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/SensorStatus'
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
          nullable: true
        equipmentId:
          type: string
          format: uuid
        equipmentName:
          type: string
          nullable: true
        sensorTypeId:
          type: string
          format: uuid
        sensorTypeName:
          type: string
          nullable: true
        thresholdId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    SensorStatus:
      enum:
        - 0
        - 1
        - 2
        - 3
      type: integer
      format: int32
...
    Unit:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
      type: integer
      format: int32
...
    UserRolesResponse:
      required:
        - roles
        - userId
      type: object
      properties:
        userId:
          type: string
          nullable: true
        roles:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
  securitySchemes:
    Bearer:
      type: http
      description: Enter the JWT token
      scheme: Bearer
      bearerFormat: JWT
...